ocr_pdf2ppt_v2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-pdf2ppt-v2"
version = "0.1.0"
description = "Robust PDF to editable PPT converter with SiliconFlow OCR"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Codex" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.5.0",
  "openai>=1.50.0",
  "pymupdf>=1.24.0",
  "python-pptx>=1.0.2",
  "Pillow>=10.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
ocr-pdf2ppt-v2 = "ocr_pdf2ppt_v2.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

